Rewrote polygon runner and line finder with state machines.
Conditional commands now run their commands and actions in the same thread as the resolver that resolved the condition. This is needed to ensure that bound commands and actions are always executed after any extra future listeners. This means that when working on the physical robot, these are currently run inside the communicator thread. More rewriting and testing is needed to ensure nothing blocks that thread for too long.

TODO: Rewrite explorer and get rid of runners.